Modified onUpdate and added onUpdate to NetSpace

+++ src/main/java/nl/jchmb/netspace/entity/manager/EntityManager.java
@@ -10,5 +10,5 @@ public interface EntityManager {
 	public Stream<Entity> stream();
 	public void addEntity(final Entity entity);
 	public void removeEntity(final Entity entity);
-	public void onUpdate(final float dt);
+	public void onUpdate(final double dt);
 }
